feat(world-protocol): modern WorldPacket/Packet framework + ConnectionType (brick A)
Widen WorldPacket opcode uint16->uint32 (default UNKNOWN_OPCODE=0xBBAADD) and
add a ConnectionType, carried through every ctor/assignment/Initialize. Add the
ConnectionType enum (REALM/INSTANCE/MAX/DEFAULT) + UNKNOWN_OPCODE to Opcodes.h;
give WorldPackets::Packet a GetConnection(), ServerPacket a connection param,
and add WorldPackets::Null. Opcode VALUES and OpcodeTable are deliberately
untouched (brick B) -- only storage/types widened so brick B's sparse 32-bit
opcodes won't truncate. Three legacy uint16(...) opcode truncations remain at
the 3.3.5 wire/DoS boundaries (commented for brick B). shared + game compile.

+++ src/server/game/Handlers/MovementHandler.cpp
@@ -343,7 +343,7 @@ void WorldSession::HandleMoveTeleportAck(WorldPacket& recvData)
 
 void WorldSession::HandleMovementOpcodes(WorldPacket& recvData)
 {
-    uint16 opcode = recvData.GetOpcode();
+    uint32 opcode = recvData.GetOpcode();
 
     Unit* mover = _player->m_mover;
